requirements.txt
django-separatedvaluesfield
